Mirror update

+++ LiteNetLib4Mirror/Assets/MirrorLib/Mirror/Runtime/Transport/Websocket/Ninja.WebSockets/Internal/WebSocketFrameReader.cs
@@ -55,7 +55,7 @@ namespace Ninja.WebSockets.Internal
             byte finBitFlag = 0x80;
             byte opCodeFlag = 0x0F;
             bool isFinBitSet = (byte1 & finBitFlag) == finBitFlag;
-            WebSocketOpCode opCode = (WebSocketOpCode) (byte1 & opCodeFlag);
+            WebSocketOpCode opCode = (WebSocketOpCode)(byte1 & opCodeFlag);
 
             // read and process second byte
             byte maskFlag = 0x80;
@@ -105,8 +105,9 @@ namespace Ninja.WebSockets.Internal
 
             if (count >= 2)
             {
-                Array.Reverse(buffer.Array, buffer.Offset, 2); // network byte order
-                int closeStatusCode = (int)BitConverter.ToUInt16(buffer.Array, buffer.Offset);
+                // network byte order
+                Array.Reverse(buffer.Array, buffer.Offset, 2);
+                int closeStatusCode = BitConverter.ToUInt16(buffer.Array, buffer.Offset);
                 if (Enum.IsDefined(typeof(WebSocketCloseStatus), closeStatusCode))
                 {
                     closeStatus = (WebSocketCloseStatus)closeStatusCode;
@@ -143,7 +144,7 @@ namespace Ninja.WebSockets.Internal
         static async Task<uint> ReadLength(byte byte2, ArraySegment<byte> smallBuffer, Stream fromStream, CancellationToken cancellationToken)
         {
             byte payloadLenFlag = 0x7F;
-            uint len = (uint) (byte2 & payloadLenFlag);
+            uint len = (uint)(byte2 & payloadLenFlag);
 
             // read a short length or a long length depending on the value of len
             if (len == 126)
@@ -152,8 +153,9 @@ namespace Ninja.WebSockets.Internal
             }
             else if (len == 127)
             {
-                len = (uint) await BinaryReaderWriter.ReadULongExactly(fromStream, false, smallBuffer, cancellationToken);
-                const uint maxLen = 2147483648; // 2GB - not part of the spec but just a precaution. Send large volumes of data in smaller frames.
+                len = (uint)await BinaryReaderWriter.ReadULongExactly(fromStream, false, smallBuffer, cancellationToken);
+                // 2GB - not part of the spec but just a precaution. Send large volumes of data in smaller frames.
+                const uint maxLen = 2147483648;
 
                 // protect ourselves against bad data
                 if (len > maxLen || len < 0)
